refactor: `src/tui/`
