+++ src/tui/display_components/__init__.py
@@ -1,0 +1,19 @@
+"""Composable display mixins for the Pitaya TUI."""
+
+from .layout import LayoutMixin
+from .render_loop import RenderLoopMixin
+from .header import HeaderMixin
+from .body import BodyMixin
+from .footer import FooterMixin
+from .snapshot import SnapshotMixin
+from .formatting import FormattingMixin
+
+__all__ = [
+    "LayoutMixin",
+    "RenderLoopMixin",
+    "HeaderMixin",
+    "BodyMixin",
+    "FooterMixin",
+    "SnapshotMixin",
+    "FormattingMixin",
+]
